+++ tests/Neo.L2.Abstractions.UnitTests/UT_NeoHubDeployReport.cs
@@ -215,6 +215,7 @@ public class UT_NeoHubDeployReport
                 + "IsForcedInclusionPipelineWiringComplete / IsSettlementClientWiringComplete / "
                 + "IsPipelineEnabled / IsSettlementPoisoned / IsSettlementIdle / IsPipelineHealthy / "
                 + "PipelineHealthFailures / IsMetricsHttpHealthy / MetricsHttpHealthFailures / "
+                + "IsLocalHostHealthy / LocalHostHealthFailures / "
                 + "HasExpectedNetwork / HasScannerDeployHeights / IsOfflinePassportComplete / "
                 + "OfflinePassportFailures / "
                 + "HasSealedBatchSink / NextExpectedBlock / ProcessCommittedBlock / IsOperatorReady / "